+++ Frontend/nubio/public/index.html
@@ -18,6 +18,11 @@
     ></script>
     <!-- 티맵 -->
     <script src="https://apis.openapi.sk.com/tmap/jsv2?version=1&appKey=prZbuvPsM53ADwzJMIxl13StkVuNvAG86O6n4YhF"></script>
+    <script
+      type="text/javascript"
+      src="//dapi.kakao.com/v2/maps/sdk.js?appkey=APIKEY&libraries=services,clusterer,drawing"
+    ></script>
+
     <title>React App</title>
   </head>
   <body style="width: 100%; height: 100%">
