+++ Frontend/nubio/src/components/common/map/KakaoMap.tsx
@@ -1,34 +1,35 @@
 // Hook
 import { useEffect, useRef, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { useLocation } from "react-router";
+// import { useDispatch, useSelector } from "react-redux";
+// import { useLocation } from "react-router";
 // 라이브러리
 import Swal from "sweetalert2";
 // 컴포넌트
-import SearchBar from "../search/SearchBar";
-import RouteInfo from "../../safeHome/route/RouteInfo";
-import ShortDirection from "../../safeHome/route/short/ShortDirection";
-import SafeDirection from "../../safeHome/route/safe/SafeDirection";
-import SelectMyLocation from "./SelectMyLocation";
+// import SearchBar from "../search/SearchBar";
+// import RouteInfo from "../../safeHome/route/RouteInfo";
+// import ShortDirection from "../../safeHome/route/short/ShortDirection";
+// import SafeDirection from "../../safeHome/route/safe/SafeDirection";
+// import SelectMyLocation from "./SelectMyLocation";
 import CalamityMessageHome from "../../safeHome/calamity/CalamityMessageHome";
 // 스타일
 import {
-  DestinationWrapper,
+  // DestinationWrapper,
   MapWrapper,
-  SearchListWrapper,
-  ClearRouteButton,
-  SearchResultsWrapper,
+  // SearchListWrapper,
+  // ClearRouteButton,
+  // SearchResultsWrapper,
 } from "../../../styles/SMap";
+import Search from "../search/Search";
 
 // redux
-import { setPosition, setTime } from "../../../redux/slice/EnjoySlice";
-import {
-  setStart,
-  setEnd,
-  setStartName,
-  setEndName,
-} from "../../../redux/slice/MapSlice";
-import { setSafeTime, setShortTime } from "../../../redux/slice/SafeSlice";
+// import { setPosition, setTime } from "../../../redux/slice/EnjoySlice";
+// import {
+// setStart,
+// setEnd,
+// setStartName,
+// setEndName,
+// } from "../../../redux/slice/MapSlice";
+// import { setSafeTime, setShortTime } from "../../../redux/slice/SafeSlice";
 
 // 카카오맵 관련
 const { kakao } = window as any;
@@ -49,59 +50,73 @@ declare global {
 }
 
 const KakaoMap = ({ position }: { position: placeItem[] }) => {
-  const dispatch = useDispatch();
   const mapRef = useRef(null);
-  const location = useLocation();
-
-  // 검색 reudx
-  const startName = useSelector(
-    (state: { map: { startName: string } }) => state.map.startName,
-  );
-  const endName = useSelector(
-    (state: { map: { endName: string } }) => state.map.endName,
-  );
-  const searchKeyword = useSelector(
-    (state: { map: { keyWord: string } }) => state.map.keyWord,
-  );
-  const safeMarkerList = useSelector(
-    (state: { safe: { safeMarkerList: Array<any> } }) =>
-      state.safe.safeMarkerList,
-  );
-  // state
   const [listIsOpen, setListIsOpen] = useState(false);
-  const [findRouteOpen, setFindRouteOpen] = useState(false);
 
-  // 마커를 담는 배열
-  let markers: any[] = [];
+  // // 검색 reudx
+  // const startName = useSelector(
+  //   (state: { map: { startName: string } }) => state.map.startName,
+  // );
+  // const endName = useSelector(
+  //   (state: { map: { endName: string } }) => state.map.endName,
+  // );
+  // const searchKeyword = useSelector(
+  //   (state: { map: { keyWord: string } }) => state.map.keyWord,
+  // );
+  // const safeMarkerList = useSelector(
+  //   (state: { safe: { safeMarkerList: Array<any> } }) =>
+  //     state.safe.safeMarkerList,
+  // );
+  // // state
+  // const [listIsOpen, setListIsOpen] = useState(false);
+  // const [findRouteOpen, setFindRouteOpen] = useState(false);
 
-  // 라인, 마커 삭제
-  const clearRoute = () => {
-    for (let j = 0; j < safeMarkerList.length; j++) {
-      safeMarkerList[j].setMap(null);
-    }
-    window.polyline?.setMap(null);
-    window.safeCustomOverlay?.setMap(null);
-    removeMarker();
-  };
-  // 맵에 표시된 경로 관련 삭제
-  const clearAllRoute = () => {
-    clearRoute();
-    window.endCustomOverlay?.setMap(null);
-    window.startCustomOverlay?.setMap(null);
-    dispatch(setStartName(""));
-    dispatch(setEndName(""));
-    dispatch(setShortTime(null));
-    dispatch(setSafeTime(null));
-    setFindRouteOpen(false);
-    removeSafeMarker();
-  };
+  // // 마커를 담는 배열
+  // let markers: any[] = [];
 
-  // 키워드 검색을 요청하는 함수
-  function searchPlaces(keyword: string) {
-    // 장소검색 객체를 통해 키워드로 장소검색을 요청
-    window.ps.keywordSearch(keyword, placesSearchCB);
-  }
+  // // 라인, 마커 삭제
+  // const clearRoute = () => {
+  //   for (let j = 0; j < safeMarkerList.length; j++) {
+  //     safeMarkerList[j].setMap(null);
+  //   }
+  //   window.polyline?.setMap(null);
+  //   window.safeCustomOverlay?.setMap(null);
+  // };
+  // // 맵에 표시된 경로 관련 삭제
+  // const clearAllRoute = () => {
+  //   removeMarker();
+  //   clearRoute();
+  //   window.endCustomOverlay?.setMap(null);
+  //   window.startCustomOverlay?.setMap(null);
+  //   dispatch(setStartName(""));
+  //   dispatch(setEndName(""));
+  //   dispatch(setShortTime(null));
+  //   dispatch(setSafeTime(null));
+  //   setFindRouteOpen(false);
+  //   removeSafeMarker();
+  // };
 
+  // // 장소검색이 완료됐을 때 호출되는 콜백함수
+  // function placesSearchCB(data: string, status: any, pagination: any) {
+  //   removeMarker();
+  //   if (status === kakao.maps.services.Status.OK) {
+  //     displayPlaces(data);
+  //     displayPagination(pagination);
+  //   } else if (status === kakao.maps.services.Status.ZERO_RESULT) {
+  //     Swal.fire({
+  //       title: "검색 결과가 존재하지 않습니다.",
+  //       text: "Nubio",
+  //     });
+  //     hideSearchResult();
+  //     return;
+  //   } else if (status === kakao.maps.services.Status.ERROR) {
+  //     Swal.fire({
+  //       title: "검색 중 오류가 발생했습니다.",
+  //       text: "Nubio",
+  //     });
+  //     return;
+  //   }
+  // }
   // 검색 결과가 없을 때는 search-result 요소를 숨기는 함수
   function hideSearchResult() {
     const searchResult = document.getElementById("search-result");
@@ -110,266 +125,226 @@ const KakaoMap = ({ position }: { position: placeItem[] }) => {
     }
   }
 
-  // 장소검색이 완료됐을 때 호출되는 콜백함수
-  function placesSearchCB(data: string, status: any, pagination: any) {
-    if (status === kakao.maps.services.Status.OK) {
-      // 정상적으로 검색이 완료됐으면
-      // 검색 목록과 마커를 표출
-      displayPlaces(data);
-      // 페이지 번호를 표출
-      displayPagination(pagination);
-    } else if (status === kakao.maps.services.Status.ZERO_RESULT) {
-      Swal.fire({
-        title: "검색 결과가 존재하지 않습니다.",
-        text: "Nubio",
-      });
-      hideSearchResult();
-      return;
-    } else if (status === kakao.maps.services.Status.ERROR) {
-      Swal.fire({
-        title: "검색 결과 중 오류가 발생했습니다.",
-        text: "Nubio",
-      });
+  // // 검색 결과 목록과 마커를 표출하는 함수
+  // function displayPlaces(places: string | any[]) {
+  //   const listEl = document.getElementById("places-list"),
+  //     resultEl = document.getElementById("search-result"),
+  //     fragment = document.createDocumentFragment(),
+  //     bounds = new window.kakao.maps.LatLngBounds();
+  //   removeMarker();
 
-      return;
-    }
-  }
+  //   listEl && removeAllChildNods(listEl);
 
-  // 검색 결과 목록과 마커를 표출하는 함수
-  function displayPlaces(places: string | any[]) {
-    const listEl = document.getElementById("places-list"),
-      resultEl = document.getElementById("search-result"),
-      fragment = document.createDocumentFragment(),
-      bounds = new window.kakao.maps.LatLngBounds();
+  //   for (var i = 0; i < places.length; i++) {
+  //     let placePosition = new window.kakao.maps.LatLng(
+  //         places[i].y,
+  //         places[i].x,
+  //       ),
+  //       marker = addMarker(placePosition, i, places[i].place_name),
+  //       itemEl = getListItem(i, places[i]);
+  //     bounds.extend(placePosition);
 
-    // 검색 결과 목록에 추가된 항목들을 제거
-    listEl && removeAllChildNods(listEl);
+  //     (function (marker, title) {
+  //       itemEl.onclick = function () {
+  //         displayInfowindow(marker, title);
+  //       };
 
-    for (var i = 0; i < places.length; i++) {
-      // 마커를 생성하고 지도에 표시
-      let placePosition = new window.kakao.maps.LatLng(
-          places[i].y,
-          places[i].x,
-        ),
-        marker = addMarker(placePosition, i, undefined),
-        itemEl = getListItem(i, places[i]); // 검색 결과 항목 Element를 생성
+  //       itemEl.onmouseout = function () {
+  //         window.infowindow.close();
+  //       };
+  //     })(marker, places[i].place_name);
 
-      bounds.extend(placePosition);
+  //     fragment.appendChild(itemEl);
+  //   }
 
-      (function (marker, title) {
-        window.kakao.maps.event.addListener(marker, "mouseover", function () {
-          displayInfowindow(marker, title);
-        });
+  //   // 검색결과 항목들을 검색결과 목록 Element에 추가
+  //   listEl && listEl.appendChild(fragment);
+  //   if (resultEl) {
+  //     resultEl.scrollTop = 0;
+  //   }
 
-        window.kakao.maps.event.addListener(marker, "mouseout", function () {
-          window.infowindow.close();
-        });
+  //   // 검색된 장소 위치를 기준으로 지도 범위를 재설정
+  //   window.map.setBounds(bounds);
+  // }
 
-        itemEl.onclick = function () {
-          displayInfowindow(marker, title);
-        };
+  // // 검색결과 항목을 Element로 반환하는 함수
+  // function getListItem(index: number, places: placeType) {
+  //   const el = document.createElement("li");
+  //   const onClickStart = () => {
+  //     dispatch(setStart({ x: places.x, y: places.y }));
+  //     dispatch(setStartName(places.place_name));
+  //     removeMarker();
+  //     window.polyline?.setMap(null);
+  //     window.startCustomOverlay?.setMap(null);
+  //     let content = `<div class ="label"  style="background:#ffc542; font-size:0.8rem; border:0.5px solid white; padding:0.3rem; border-radius:1rem; color:white;"></span><span class="center">
+  //     출발</span><span class="right"></span></div>`;
+  //     let markerPosition = new window.kakao.maps.LatLng(places.y, places.x);
+  //     let customOverlay = new window.kakao.maps.CustomOverlay({
+  //       position: markerPosition,
+  //       content: content,
+  //     });
+  //     window.startCustomOverlay = customOverlay;
+  //     window.startCustomOverlay.setMap(window.map);
+  //   };
+  //   const onClickEnd = () => {
+  //     dispatch(setEnd({ x: places.x, y: places.y }));
+  //     dispatch(setEndName(places.place_name));
+  //     removeMarker();
+  //     window.polyline?.setMap(null);
+  //     window.endCustomOverlay?.setMap(null);
 
-        itemEl.onmouseout = function () {
-          window.infowindow.close();
-        };
-      })(marker, places[i].place_name);
+  //     let content = `<div class ="label" style="background:#f9373f; font-size:0.8rem; border:0.5px solid white; padding:0.3rem; border-radius:1rem; color:white;"><span class="left"></span><span class="center">도착</span><span class="right"></span></div>`;
+  //     let markerPosition = new window.kakao.maps.LatLng(places.y, places.x);
+  //     let customOverlay = new window.kakao.maps.CustomOverlay({
+  //       position: markerPosition,
+  //       content: content,
+  //     });
 
-      fragment.appendChild(itemEl);
-    }
+  //     window.endCustomOverlay = customOverlay;
+  //     window.endCustomOverlay.setMap(window.map);
+  //   };
+  //   const ClickPlace = (place: placeType) => {
+  //     dispatch(setPosition(place));
+  //   };
+  //   if (places.length !== 0) {
+  //     const isEnjoy = location.pathname.includes("/enjoy");
+  //     let itemStr = `
+  //       <div class="info">
+  //       <div class="name">
+  //           <h3 class="info-item place-name">${index + 1}. ${
+  //       places.place_name
+  //     }<a id="homePage" href=${places.place_url}>상세보기</a>
+  //     </h3>
 
-    // 검색결과 항목들을 검색결과 목록 Element에 추가
-    listEl && listEl.appendChild(fragment);
-    if (resultEl) {
-      resultEl.scrollTop = 0;
-    }
+  //   </div>
+  //           ${
+  //             places.road_address_name
+  //               ? `<span class="address ">
+  //                 ${places.road_address_name}
+  //                   ${places.address_name}
+  //                   </span>`
+  //               : `<span class="address ">
+  //                   ${places.address_name}
+  //               </span>`
+  //           }
+  //           <span class="tel">
+  //             ${places.phone}
+  //           </span>
+  //           <div class="bnt">
+  //             <button id="start">출발</button>
+  //             <button id="end">도착</button>
+  //           </div>
+  //         </a>
+  //       </div>
+  //       `;
+  //     el.innerHTML = itemStr;
+  //     el.className = "item";
+  //     const startButton = el.querySelector("#start");
+  //     startButton?.addEventListener("click", onClickStart);
+  //     const endButton = el.querySelector("#end");
+  //     endButton?.addEventListener("click", onClickEnd);
+  //     if (isEnjoy) {
+  //       const select = el.querySelector(".place-name");
+  //       select?.addEventListener("click", () => ClickPlace(places));
+  //     }
+  //   }
 
-    // 검색된 장소 위치를 기준으로 지도 범위를 재설정
-    window.map.setBounds(bounds);
-  }
+  //   return el;
+  // }
+  // // 마커를 생성하고 지도 위에 마커를 표시하는 함수
+  // function addMarker(position: any, idx: number, title: string) {
+  //   var imageSrc =
+  //       "https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/marker_number_blue.png",
+  //     imageSize = new window.kakao.maps.Size(36, 37),
+  //     imgOptions = {
+  //       spriteSize: new window.kakao.maps.Size(36, 691),
+  //       spriteOrigin: new window.kakao.maps.Point(0, idx * 46 + 10),
+  //       offset: new window.kakao.maps.Point(13, 37),
+  //     },
+  //     markerImage = new window.kakao.maps.MarkerImage(
+  //       imageSrc,
+  //       imageSize,
+  //       imgOptions,
+  //     );
+  //   const marker = new kakao.maps.Marker({
+  //     position: position,
+  //     image: markerImage,
+  //   });
 
-  // 검색결과 항목을 Element로 반환하는 함수
-  function getListItem(index: number, places: placeType) {
-    const el = document.createElement("li");
-    const onClickStart = () => {
-      dispatch(setStart({ x: places.x, y: places.y }));
-      dispatch(setStartName(places.place_name));
-      removeMarker();
-      window.polyline?.setMap(null);
-      window.startCustomOverlay?.setMap(null);
+  //   marker.setMap(window.map);
+  //   markers.push(marker);
 
-      let content = `<div class ="label"  style="background:#ffc542; font-size:0.8rem; border:0.5px solid white; padding:0.3rem; border-radius:1rem; color:white;"></span><span class="center">
-      출발</span><span class="right"></span></div>`;
-      // 커스텀 오버레이가 표시될 위치입니다
-      let markerPosition = new window.kakao.maps.LatLng(places.y, places.x);
-      // 커스텀 오버레이를 생성합니다
-      let customOverlay = new window.kakao.maps.CustomOverlay({
-        position: markerPosition,
-        content: content,
-      });
-      window.startCustomOverlay = customOverlay;
-      // 커스텀 오버레이를 지도에 표시합니다
-      window.startCustomOverlay.setMap(window.map);
-    };
-    const onClickEnd = () => {
-      dispatch(setEnd({ x: places.x, y: places.y }));
-      dispatch(setEndName(places.place_name));
-      removeMarker();
-      window.polyline?.setMap(null);
-      window.endCustomOverlay?.setMap(null);
+  //   return marker;
+  // }
 
-      let content = `<div class ="label" style="background:#f9373f; font-size:0.8rem; border:0.5px solid white; padding:0.3rem; border-radius:1rem; color:white;"><span class="left"></span><span class="center">도착</span><span class="right"></span></div>`;
-      let markerPosition = new window.kakao.maps.LatLng(places.y, places.x);
-      let customOverlay = new window.kakao.maps.CustomOverlay({
-        position: markerPosition,
-        content: content,
-      });
+  // // 지도 위에 표시되고 있는 마커를 모두 제거합니다
+  // const removeMarker = () => {
+  //   for (var i = 0; i < markers.length; i++) {
+  //     markers[i].setMap(null);
+  //   }
+  //   markers = [];
+  // };
+  // const safePlaces = useSelector(
+  //   (state: { safe: { safePlace: any } }) => state.safe.safePlace,
+  // );
+  // const removeSafeMarker = () => {
+  //   for (let i = 0; i < safePlaces.length; i++) {
+  //     window.safeCustomOverlay.setMap(null);
+  //   }
+  // };
+  // // 검색결과 목록 하단에 페이지번호를 표시는 함수
+  // function displayPagination(pagination: {
+  //   last: number;
+  //   current: number;
+  //   gotoPage: (arg0: number) => void;
+  // }) {
+  //   const paginationEl = document.getElementById("pagination") as HTMLElement;
+  //   let fragment = document.createDocumentFragment();
+  //   let i;
 
-      window.endCustomOverlay = customOverlay;
-      window.endCustomOverlay.setMap(window.map);
-    };
-    const ClickPlace = (place: placeType) => {
-      dispatch(setPosition(place));
-    };
-    if (places.length !== 0) {
-      const isEnjoy = location.pathname.includes("/enjoy");
-      let itemStr = `
-        <div class="info">
-        <div class="name">
-            <h3 class="info-item place-name">${index + 1}. ${
-        places.place_name
-      }<a id="homePage" href=${places.place_url}>상세보기</a>
-      </h3>
-        
-    </div>
-            ${
-              places.road_address_name
-                ? `<span class="address ">
-                  ${places.road_address_name}
-                    ${places.address_name}
-                    </span>`
-                : `<span class="address ">
-                    ${places.address_name}
-                </span>`
-            }
-            <span class="tel">
-              ${places.phone}
-            </span>
-            <div class="bnt">
-              <button id="start">출발</button>
-              <button id="end">도착</button>
-            </div>
-          </a>
-        </div>
-        `;
-      el.innerHTML = itemStr;
-      el.className = "item";
-      const startButton = el.querySelector("#start");
-      startButton?.addEventListener("click", onClickStart);
-      const endButton = el.querySelector("#end");
-      endButton?.addEventListener("click", onClickEnd);
-      if (isEnjoy) {
-        const select = el.querySelector(".place-name");
-        select?.addEventListener("click", () => ClickPlace(places));
-      }
-    }
+  //   while (paginationEl.hasChildNodes()) {
+  //     paginationEl.lastChild &&
+  //       paginationEl.removeChild(paginationEl.lastChild);
+  //   }
 
-    return el;
-  }
-  // 마커를 생성하고 지도 위에 마커를 표시하는 함수
-  function addMarker(position: any, idx: number, title: undefined) {
-    var imageSrc =
-        "https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/marker_number_blue.png", // 마커 이미지 url, 스프라이트 이미지
-      imageSize = new window.kakao.maps.Size(36, 37), // 마커 이미지의 크기
-      imgOptions = {
-        spriteSize: new window.kakao.maps.Size(36, 691), // 스프라이트 이미지의 크기
-        spriteOrigin: new window.kakao.maps.Point(0, idx * 46 + 10), // 스프라이트 이미지 중 사용할 영역의 좌상단 좌표
-        offset: new window.kakao.maps.Point(13, 37), // 마커 좌표에 일치시킬 이미지 내에서의 좌표
-      },
-      markerImage = new window.kakao.maps.MarkerImage(
-        imageSrc,
-        imageSize,
-        imgOptions,
-      ),
-      marker = new window.kakao.maps.Marker({
-        position: position, // 마커의 위치
-        image: markerImage,
-      });
+  //   for (i = 1; i <= pagination.last; i++) {
+  //     const el = document.createElement("a") as HTMLAnchorElement;
+  //     el.href = "#";
+  //     el.innerHTML = i.toString();
 
-    marker.setMap(window.map); // 지도 위에 마커를 표출
-    markers.push(marker); // 배열에 생성된 마커를 추가
+  //     if (i === pagination.current) {
+  //       el.className = "on";
+  //     } else {
+  //       el.onclick = (function (i) {
+  //         return function () {
+  //           pagination.gotoPage(i);
+  //         };
+  //       })(i);
+  //     }
 
-    return marker;
-  }
+  //     fragment.appendChild(el);
+  //   }
+  //   paginationEl.appendChild(fragment);
+  // }
 
-  // 지도 위에 표시되고 있는 마커를 모두 제거합니다
-  const removeMarker = () => {
-    for (var i = 0; i < markers.length; i++) {
-      markers[i].setMap(null);
-    }
-    markers = [];
-  };
-  const safePlaces = useSelector(
-    (state: { safe: { safePlace: any } }) => state.safe.safePlace,
-  );
-  const removeSafeMarker = () => {
-    for (let i = 0; i < safePlaces.length; i++) {
-      window.safeCustomOverlay.setMap(null);
-    }
-  };
-  // 검색결과 목록 하단에 페이지번호를 표시는 함수
-  function displayPagination(pagination: {
-    last: number;
-    current: number;
-    gotoPage: (arg0: number) => void;
-  }) {
-    const paginationEl = document.getElementById("pagination") as HTMLElement;
-    let fragment = document.createDocumentFragment();
-    let i;
+  // // 검색결과 목록 또는 마커를 클릭했을 때 호출되는 함수
+  // // 인포윈도우에 장소명을 표시
+  // function displayInfowindow(marker: any, title: string) {
+  //   const content =
+  //     '<div style="padding:5px;z-index:1;" class="marker-title">' +
+  //     title +
+  //     "</div>";
 
-    // 기존에 추가된 페이지번호를 삭제
-    while (paginationEl.hasChildNodes()) {
-      paginationEl.lastChild &&
-        paginationEl.removeChild(paginationEl.lastChild);
-    }
+  //   window.infowindow.setContent(content);
+  //   window.infowindow.open(window.map, marker);
+  // }
 
-    for (i = 1; i <= pagination.last; i++) {
-      const el = document.createElement("a") as HTMLAnchorElement;
-      el.href = "#";
-      el.innerHTML = i.toString();
-
-      if (i === pagination.current) {
-        el.className = "on";
-      } else {
-        el.onclick = (function (i) {
-          return function () {
-            pagination.gotoPage(i);
-          };
-        })(i);
-      }
-
-      fragment.appendChild(el);
-    }
-    paginationEl.appendChild(fragment);
-  }
-
-  // 검색결과 목록 또는 마커를 클릭했을 때 호출되는 함수
-  // 인포윈도우에 장소명을 표시
-  function displayInfowindow(marker: any, title: string) {
-    const content =
-      '<div style="padding:5px;z-index:1;" class="marker-title">' +
-      title +
-      "</div>";
-
-    window.infowindow.setContent(content);
-    window.infowindow.open(window.map, marker);
-  }
-
-  // 검색결과 목록의 자식 Element를 제거하는 함수
-  function removeAllChildNods(el: HTMLElement) {
-    while (el.hasChildNodes()) {
-      el.lastChild && el.removeChild(el.lastChild);
-    }
-  }
+  // // 검색결과 목록의 자식 Element를 제거하는 함수
+  // function removeAllChildNods(el: HTMLElement) {
+  //   while (el.hasChildNodes()) {
+  //     el.lastChild && el.removeChild(el.lastChild);
+  //   }
+  // }
   // 현재위치
   const startCurPosition = () => {
     if (navigator.geolocation) {
@@ -380,21 +355,19 @@ const KakaoMap = ({ position }: { position: placeItem[] }) => {
           window.map.setCenter(
             new window.kakao.maps.LatLng(window.myLatitude, window.myLongitude),
           );
-          // 현재 위치에 마커를 표시
           const marker = new kakao.maps.Marker({
             position: new kakao.maps.LatLng(
               window.myLatitude,
               window.myLongitude,
             ),
           });
-          marker.setMap(window.map); // 마커를 지도에 표시
+          marker.setMap(window.map);
         },
         (error) => {
           Swal.fire({
             title: `geolocation ${error} 발생.`,
             text: "NUBIO",
           });
-          // console.error("geolocation 에러 발생:", error);
         },
       );
     } else {
@@ -402,7 +375,6 @@ const KakaoMap = ({ position }: { position: placeItem[] }) => {
         title: "지금 브라우저에서는 geolocation를 지원하지 않습니다.",
         text: "NUBIO",
       });
-      // console.error("지금 브라우저에서는 geolocation를 지원하지 않습니다.");
     }
   };
 
@@ -414,20 +386,16 @@ const KakaoMap = ({ position }: { position: placeItem[] }) => {
       center: new kakao.maps.LatLng(0, 0),
       level: 3, // 지도의 확대 레벨
     };
-    // mapRef가 초기화되지 않았다면 지도 인스턴스를 생성하고 참조합니다.
     if (!mapRef.current) {
       const map = new kakao.maps.Map(mapContainer, mapOption);
       mapRef.current = map;
       window.map = map;
-      // 장소 검색 객체를 생성
       const ps = new kakao.maps.services.Places();
       window.ps = ps;
 
-      // 검색 결과 목록이나 마커를 클릭했을 때 장소명을 표출할 인포윈도우를 생성
       const infowindow = new kakao.maps.InfoWindow({ zIndex: 1 });
       window.infowindow = infowindow;
 
-      // 맵 클릭 시 검색리스트 안보임
       window.map.addListener("click", () => {
         setListIsOpen(false);
       });
@@ -436,8 +404,9 @@ const KakaoMap = ({ position }: { position: placeItem[] }) => {
   return (
     <>
       <MapWrapper id="map" className="map" />
-      <CalamityMessageHome />
-      <SearchBar
+      <Search />
+      {/* <CalamityMessageHome /> */}
+      {/* <SearchBar
         searchPlaces={searchPlaces}
         setListIsOpen={setListIsOpen}
         setFindRouteOpen={setFindRouteOpen}
@@ -470,7 +439,7 @@ const KakaoMap = ({ position }: { position: placeItem[] }) => {
           </SearchListWrapper>
           <div id="pagination"></div>
         </SearchResultsWrapper>
-      ) : null}
+      ) : null} */}
     </>
   );
 };
